Move button to cta folder + fix inputs (regression) + lint
