+++ packages/native/src/components/Form/Input/SearchInput/index.tsx
@@ -4,7 +4,7 @@ import Input, {
   InputRenderLeftContainer,
 } from "@components/Form/Input/BaseInput";
 import SearchMedium from "@ui/assets/icons/SearchMedium";
-import styled from "styled-components";
+import styled from "styled-components/native";
 
 const Icon = styled(SearchMedium).attrs((p) => ({
   color: p.theme.colors.palette.neutral.c70,
