+++ packages/native/storybook/stories/CenterView/index.tsx
@@ -8,16 +8,16 @@ const Main = styled.View`
   flex: 1;
   justify-content: center;
   align-items: center;
-  background-color: ${p => p.theme.colors.palette.neutral.c00};
+  background-color: ${(p) => p.theme.colors.palette.neutral.c00};
   overflow: scroll;
 `;
 
 const ThemeButton = styled.TouchableOpacity`
-  width: ${p => p.theme.space[6]}px;
-  height: ${p => p.theme.space[6]}px;
-  border-top-left-radius: ${p => p.theme.space[6]};
-  border-bottom-left-radius: ${p => p.theme.space[6]};
-  background-color: ${p => p.theme.colors.palette.neutral.c100};
+  width: ${(p) => p.theme.space[6]}px;
+  height: ${(p) => p.theme.space[6]}px;
+  border-top-left-radius: ${(p) => p.theme.space[6]};
+  border-bottom-left-radius: ${(p) => p.theme.space[6]};
+  background-color: ${(p) => p.theme.colors.palette.neutral.c100};
   justify-content: center;
   align-items: center;
   position: absolute;
@@ -27,7 +27,7 @@ const ThemeButton = styled.TouchableOpacity`
 `;
 
 const Icon = styled.Text`
-  color: ${p => p.theme.colors.palette.neutral.c100};
+  color: ${(p) => p.theme.colors.palette.neutral.c100};
   font-size: 15px;
 `;
